Feat : InputForm 컴포넌트화 #62
